requirements.txt
streamlit
python-dotenv
PyGithub
streamlit-agraph
google-generativeai